rewrite-complete
