--- apps/ai/eslint.config.mjs
@@ -1,3 +1,0 @@
-import { config } from "@roro-ai/eslint-config/base"
-
-export default config
