--- apps/ai/package.json
@@ -1,64 +1,0 @@
-{
-  "name": "ai",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "dev": "tsx --watch src/index.ts",
-    "start": "node dist/index.js",
-    "build": "tsc",
-    "type-check": "tsc --noEmit",
-    "lint": "eslint .",
-    "lint:fix": "eslint --fix .",
-    "format": "prettier --write .",
-    "format:check": "prettier --check ."
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@deepgram/sdk": "^3.9.0",
-    "@discordjs/opus": "github:discordjs/opus",
-    "@google-cloud/text-to-speech": "^5.8.0",
-    "@google/generative-ai": "^0.21.0",
-    "@grpc/grpc-js": "^1.12.6",
-    "@langchain/community": "^0.3.30",
-    "@langchain/core": "^0.3.39",
-    "@langchain/google-genai": "^0.1.8",
-    "@langchain/groq": "^0.1.3",
-    "@langchain/langgraph": "^0.2.46",
-    "@roro-ai/database": "workspace:*",
-    "@socket.io/redis-adapter": "^8.3.0",
-    "@types/ws": "^8.5.14",
-    "chromadb": "^1.10.4",
-    "cors": "^2.8.5",
-    "dotenv": "^16.4.7",
-    "elevenlabs": "^1.51.0",
-    "express": "^4.21.2",
-    "fluent-ffmpeg": "^2.1.3",
-    "langchain": "^0.3.15",
-    "mediasoup": "^3.15.3",
-    "node-pre-gyp": "^0.17.0",
-    "opus-decoder": "^0.7.7",
-    "opusscript": "^0.1.1",
-    "playht": "^0.16.0",
-    "prism-media": "^1.3.5",
-    "rebuild": "^0.1.2",
-    "redis": "^4.7.0",
-    "rtp-parser": "^1.0.0",
-    "rtp.js": "^0.13.1",
-    "socket.io": "^4.8.1",
-    "sox-stream": "^2.0.6",
-    "ws": "^8.18.0"
-  },
-  "devDependencies": {
-    "@roro-ai/eslint-config": "workspace:*",
-    "@roro-ai/typescript-config": "workspace:*",
-    "@types/cors": "^2.8.17",
-    "@types/express": "^5.0.0",
-    "@types/node": "^22.13.1",
-    "@types/rtp-parser": "^1.0.4",
-    "typescript": "^5.7.3"
-  }
-}
