--- apps/ai/src/event/index.ts
@@ -1,3 +1,0 @@
-import EventEmitter from "events";
-
-export const event = new EventEmitter()
