--- apps/ai/src/index.ts
@@ -1,28 +1,0 @@
-import { createServer } from "node:http";
-import { app } from "./module/app.js";
-import 'dotenv/config'
-import { mediasoupWorkerManager } from "./managers/worker-manager.js";
-import { AiSocketServer } from "./classes/ai-socket-server.js";
-
-
-
-const port = process.env.PORT || 3333;
-export const server = createServer(app);
-
-(() => {
-  try {
-    const aiSocketServer = AiSocketServer.getInstance()
-    aiSocketServer.initialize();
-    mediasoupWorkerManager.createWorkers();
-  } catch (error) {
-    console.error("Error during initialization:", error);
-  }
-})()
-
-server.listen(port, () => {
-  console.log(`Listening at http://localhost:${port}`);
-});
-server.on('error', console.error);
-
-
-
