--- apps/ai/src/module/app.ts
@@ -1,16 +1,0 @@
-import express, {type Application } from "express"
-import cors from "cors"
-
-export const app: Application = express();
-
-app.use(cors({
-    origin: "http://localhost:3000",
-    methods: ["GET" , "POST", "DELETE", "PUT"],
-    credentials: true,
-}))
-
-app.use(express.json())
-
-
-
-
