--- apps/ai/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": ["@roro-ai/typescript-config/base.json"],
-  "exclude": ["node_modules", "dist"],
-  "include": ["src/**/*.ts"],
-  "compilerOptions": {
-    "noEmit": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "typeRoots": ["node_modules/@types"],
-  }
-}
