+++ apps/client/src/components/room/join.tsx
@@ -1,24 +1,18 @@
 import { useSession } from '@/lib/auth-client'
 import { useMediaStore } from '@/store/useMedia'
-import { useMediasoupConsumerStore } from '@/store/useMediasoupConsumerStore'
-import { useMediasoupProducerStore } from '@/store/useMediasoupProducerStore'
+
+import { useMediasoupStore } from '@/store/useMediasoupStore'
 import { Button } from '@roro-ai/ui/components/ui/button'
 import { useParams } from 'next/navigation'
 import React from 'react'
 
 export default function Join() {
-      const joinRoom = useMediasoupProducerStore((state) => state.joinRoom)
-      const createSendTransport = useMediasoupProducerStore((state) => state.createSendTransport)
-      const createProducerPlainTransport = useMediasoupProducerStore((state) => state.createProducerPlainTransport)
-      const setupDevice = useMediasoupProducerStore((state) => state.setDevice)
-      const startProducing = useMediasoupProducerStore((state) => state.startProducing)
-      const consumerJoinRoom = useMediasoupConsumerStore((state) => state.consumerJoinRoom)
-      const createRecvTransport = useMediasoupConsumerStore((state) => state.createRecvTransport)
-      const createConsumerPlainTransport = useMediasoupConsumerStore((state) => state.createConsumerPlainTransport)
-      const connectProducerPlainTransport = useMediasoupProducerStore((state) => state.connectProducerPlainTransport)
-      const startConsumingProducer = useMediasoupConsumerStore((state) => state.startConsumerProducing)
-      const startConsuming = useMediasoupConsumerStore((state) => state.startConsuming)
-      const forwardMedia = useMediasoupProducerStore((state) => state.forwardMedia)
+      const joinRoom = useMediasoupStore((state) => state.joinRoom)
+      const createSendTransport = useMediasoupStore((state) => state.createSendTransport)
+      const setupDevice = useMediasoupStore((state) => state.setDevice)
+      const startProducing = useMediasoupStore((state) => state.startProducing)
+      const createRecvTransport = useMediasoupStore((state) => state.createRecvTransport)
+      const startConsuming = useMediasoupStore((state) => state.startConsuming)
       const localStream = useMediaStore((state) => state.localStream)
       const params = useParams()
       const {data:session} = useSession()
@@ -35,21 +29,13 @@ export default function Join() {
 
       const JoinHandler = async () => {
         await joinRoom(roomId, userId, username)
-        await consumerJoinRoom(roomId)
         const device = await setupDevice()
         const response = await createRecvTransport(roomId, device)
         if (response.success) {
           await createSendTransport(roomId)
-        await createProducerPlainTransport(roomId)
-        const plainTransportParams = await createConsumerPlainTransport(roomId)
-        await connectProducerPlainTransport(plainTransportParams, roomId)
     
-        if (!localStream) return;
-
+        if (!localStream) return
         await startProducing(localStream)
-        const rtpParameters = await forwardMedia(roomId)
-        console.log("forward-2", rtpParameters)
-        await startConsumingProducer(roomId,rtpParameters)
         await startConsuming(device, roomId)
         }
