+++ apps/client/src/store/useMediasoupStore.ts
@@ -13,24 +13,22 @@ type MediasoupProducerState = {
     producers: {
       audio: mediasoupClient.types.Producer | null
     }
+    recvTransport: mediasoupClient.types.Transport | null
+    remoteStream: MediaStream | null
 
 
 
     setDevice: () => Promise<mediasoupClient.types.Device>
     joinRoom: (roomId: string, userId: string, username: string) => Promise<void>
     getDevice: () => Promise<mediasoupClient.types.Device>
-    forwardMedia: (roomId: string) => Promise<mediasoupClient.types.RtpParameters>
     createSendTransport: (roomId: string) => Promise<void>
-    createProducerPlainTransport: (roomId: string) => Promise<{ip: string, port: number, rtcpPort: number }>
     startProducing: (localstream: MediaStream) => Promise<void>
-    connectProducerPlainTransport: (
-            plainTransportParams: { ip: string; port: number; rtcpPort?: number },
-            roomId: string
-    ) => Promise<{ success: boolean; error?: string }>;
+    createRecvTransport: (roomId: string, device: mediasoupClient.types.Device) => Promise<{ success: boolean; error?: string }>;
+    startConsuming: (device: mediasoupClient.types.Device, roomId:string) => Promise<void>
     
 }
 
-export const useMediasoupProducerStore = create<MediasoupProducerState>((set, get) => ({
+export const useMediasoupStore = create<MediasoupProducerState>((set, get) => ({
     device: null,
     routerRtpCapabilities: null,
     joined: false,
@@ -38,7 +36,10 @@ export const useMediasoupProducerStore = create<MediasoupProducerState>((set, ge
     error: "",
     producers: {
       audio: null,
-  },
+    },
+    remoteStream: null,
+    recvTransport: null,
+
   
 
 
@@ -102,6 +103,71 @@ export const useMediasoupProducerStore = create<MediasoupProducerState>((set, ge
                 set({ error: "Failed to start media "})
                 throw new Error(`[Failed to start stream]: ${(error as Error).message}`);
               }
+    },
+
+    startConsuming: async (device, roomId) => {
+   
+      const { recvTransport } = get();
+      if (!recvTransport) {
+        throw new Error('Receive transport not initialized');
+      }
+      if (!device || !device.loaded) {
+        throw new Error("Mediasoup device is not loaded.");
+      }
+      try {
+        
+        const response = await socket.emitWithAck('consume-media', {
+          roomId,
+          rtpCapabilities: device.rtpCapabilities,
+        });
+        const { consumerParams } = response;
+        
+
+        if (typeof consumerParams === "string") {
+          if (consumerParams === "noProducer") {
+            console.log("There is no producer set up to consume");
+            throw new Error("No producer available");
+          } else if (consumerParams === "cannotConsume") {
+            console.log("rtpCapabilities failed. Cannot consume");
+            throw new Error("Cannot consume producer");
+          }
+        }
+        
+        console.log("Consumer params:", consumerParams);
+        
+        
+        const consumer = await recvTransport.consume({
+          producerId: consumerParams.producerId,
+          id: consumerParams.id,
+          kind: consumerParams.kind,
+          rtpParameters: consumerParams.rtpParameters,
+        });
+
+        console.log("🎧 Consumer created:", consumer);
+        
+        
+        consumer.track.addEventListener("ended", () => {
+          console.log("Remote track has ended");
+        });
+        consumer.track.onmute = () => {
+          console.log("Remote track muted");
+        };
+        consumer.track.onunmute = () => {
+          console.log("Remote track unmuted");
+        };
+
+        const newRemoteStream = new MediaStream([consumer.track]);
+        
+        await socket.emitWithAck("unpauseConsumer", { roomId });
+        
+        
+        consumer.resume()
+
+        set({ remoteStream: newRemoteStream });
+        
+      } catch (error) {
+        throw new Error(`Failed to start consuming: ${(error as Error).message}`);
+      }
     },
 
     createSendTransport: async(roomId) => {
@@ -169,64 +235,68 @@ export const useMediasoupProducerStore = create<MediasoupProducerState>((set, ge
           }
     },
 
-    connectProducerPlainTransport: async (plainTransportParams, roomId) => {
-            try {
-                if (!socket?.connected) {
-                    throw new Error("AI socket is not connected");
-                }
-    
-                console.log(plainTransportParams)
-    
-                const response = await socket.emitWithAck("connect-plain-transport", {
-                  ip: plainTransportParams.ip,
-                  port: plainTransportParams.port,
-                  rtcpPort: plainTransportParams.rtcpPort,
-                  roomId
-                })
-          
-                if(response.success !== true) {
-                    set({ error: "Failed to connect plain transport. Please try again later." });
-                    return { success: false, error: "Failed to connect plain transport." };
-                }
-    
-                return { success: true };
-          
-              } catch (error) {
-    
-                console.error("Error connecting plain transport:", error);
-                set({ error: "Error connecting plain transport. Please check your connection." });
-                return { success: false, error: (error as Error).message };
+    createRecvTransport: async (roomId, device) => {
+          try {
+              if (!socket || !socket.connected) {
+                  throw new Error("AI socket is not connected");
               }
+      
+              const { clientTransportParams } = await socket.emitWithAck('createConsumerTransport', { roomId });
+      
+              if (!clientTransportParams || !clientTransportParams.dtlsParameters) {
+                  throw new Error("Invalid transport parameters received");
+              }
+      
+              console.log("🚚 client transport params", clientTransportParams.dtlsParameters);
+      
+              const transport = device.createRecvTransport(clientTransportParams);
+    
+              console.log("🚚 Transport Consumer:", transport)
+      
+              console.log("🚚 Consumer transport", transport);
+      
+              transport.on('connectionstatechange', (state) => {
+                  console.log('Consumer transport connection state change:', state);
+              });
+      
+              transport.on('icegatheringstatechange', (state) => {
+                  console.log('Consumer transport ICE gathering state change:', state);
+              });
+      
+              transport.on('connect', async ({ dtlsParameters }, callback, errback) => {
+                  try {
+                      const response = await socket.emitWithAck('connect-consumer-transport', {
+                          roomId,
+                          dtlsParameters,
+                      });
+      
+                      if (response.success) {
+                          callback();
+                      } else {
+                          throw new Error('Failed to connect consumer transport on server side');
+                      }
+                  } catch (error) {
+                      console.error('Error connecting consumer transport:', error);
+                      errback(error instanceof Error ? error : new Error("Unknown error occurred"));
+                  }
+              });
+      
+              if (typeof set === "function") {
+                  set({ recvTransport: transport });
+              } else {
+                  console.warn("set function is undefined, transport not stored.");
+              }
+      
+              return { success: true };
+      
+          } catch (error) {
+              console.error("Error creating receive transport:", error);
+              if (typeof set === "function") {
+                  set({ error: "Error creating receive transport. Please try again later." });
+              }
+              return { success: false, error: error instanceof Error ? error.message : "Unknown error occurred" };
+          }
       },
-
-    forwardMedia: async (roomId) => {
-           try {
-             const rtpParameters = await socket.emitWithAck("forward-media", { roomId })
-
-             console.log("Forward", rtpParameters)
-
-             return rtpParameters
-           } catch (error) {
-            console.error(`Failed to forward media: ${(error as Error).message}`);
-            set({ error: "Failed to process the request please try again later"})
-            throw new Error(`Failed to forward media: ${(error as Error).message}`);
-           }
-    },
-
-    createProducerPlainTransport: async (roomId) => {
-        try {
-            const plainTransportParams = await socket.emitWithAck("create-plain-transport", {
-                roomId
-            })
-
-            console.log("plain", plainTransportParams)
-
-            return plainTransportParams
-        } catch (error) {
-            console.error("Failed to create plain Transport", error)
-            set({ error: "Something went wrong please try again later"})
-        }
-    },
 
     getDevice: async () => {
        try {
