+++ apps/media/package.json
@@ -26,8 +26,10 @@
     "fluent-ffmpeg": "^2.1.3",
     "langchain": "^0.3.15",
     "mediasoup": "^3.15.3",
+    "opusscript": "^0.1.1",
     "rtp-ogg-opus": "^1.0.13",
     "rtp-parser": "^1.0.0",
+    "rtp.js": "^0.13.1",
     "socket.io": "^4.8.1",
     "sox-stream": "^2.0.6",
     "ws": "^8.18.0"
